Refactor model tests into separate modules
