+++ backend/inventory/tests/models/__init__.py
@@ -1,0 +1,1 @@
+"""Model-specific test modules."""
