--- backend/inventory/tests/test_models.py
@@ -1,127 +1,0 @@
-import pytest
-import random
-import string
-from django.core.exceptions import ValidationError
-from django.contrib.auth.models import User
-from inventory.models import UserProfile, Household
-from inventory.tests.factories import factories
-
-# User / UserProfile Model
-
-@pytest.mark.django_db
-def test_user_profile_created_on_user_creation():
-    user = factories.UserFactory()
-    # user = User.objects.create_user(username="testuser", password="testpass")
-
-    # Try to get the related user profile
-    try:
-        profile = UserProfile.objects.get(user=user)
-    except UserProfile.DoesNotExist:
-        pytest.fail("UserProfile was not created for the new user")
-
-
-@pytest.mark.django_db
-def test_user_creation_requires_username():
-    with pytest.raises(ValueError, match="must be set"):
-        User.objects.create_user(username="", password="testpass")
-
-
-@pytest.mark.django_db
-def test_user_creation_requires_password(): 
-    user = User.objects.create_user(username="testuser", password=None)
-    assert user.has_usable_password() is False
-
-
-@pytest.mark.django_db
-def test_updated_at_changes_on_profile_update():
-    user = User.objects.create_user(username="timestamptest", password="abc123")
-    profile = user.userprofile
-
-    original_updated_at = profile.updated_at
-
-    import time;
-    time.sleep(1)
-
-    profile.zip_code = "90083"
-    profile.save()
-
-    profile.refresh_from_db()
-    assert profile.updated_at > original_updated_at
-
-
-@pytest.mark.django_db
-def test_user_profile_assigned_to_household():
-    user = User.objects.create_user(username='testuser', password="abc123")
-    household = Household.objects.create(name="Test Household")
-
-    profile = user.userprofile
-    profile.household = household
-    profile.save()
-
-    assert profile.household is not None
-    assert profile.household.name == "Test Household"
-    assert household.members.filter(user=user).exists()
-
-# Household Model Tests
-
-# Inventory Model Tests
-# -inventory assigned to household, when household gets deleted the inventor household_id is set to null
-
-
-
-@pytest.mark.django_db
-def test_inventory_assigned_to_household():
-    inventory = factories.InventoryFactory()
-    household = factories.HouseholdFactory(name='Test Household')
-
-    inventory.household = household
-    inventory.save()
-
-    assert inventory.household is not None
-    assert inventory.household.name == 'Test Household'
-    assert household.id == inventory.household.id
-
-@pytest.mark.django_db
-def test_inventory_timestamp_updates_after_edits():
-    inventory = factories.InventoryFactory()
-
-    first_update_time = inventory.updated_at
-    initial_creation_time = inventory.created_at
-
-    inventory.name = 'New Last Name'
-    inventory.save()
-
-    assert initial_creation_time == inventory.created_at
-    assert first_update_time < inventory.updated_at
-     
-@pytest.mark.django_db
-def test_inventory_household_set_to_null_upon_household_deletion():
-    inventory = factories.InventoryFactory()
-    household = factories.HouseholdFactory()
-
-    inventory.household = household
-    inventory.save()
-
-    assert inventory.household is not None
-    assert household.id == inventory.household.id
-
-    inventory.household.delete()
-    inventory.refresh_from_db()
-
-    assert inventory.household is None
-
-@pytest.mark.django_db
-def test_inventory_name_length_constraint():
-    inventory = factories.InventoryFactory()
-
-    too_long_string = ''.join(random.choices(string.ascii_letters + string.digits, k=101))
-    inventory.name = too_long_string
-
-    with pytest.raises(ValidationError) as excinfo:
-        inventory.full_clean()
-
-        assert (
-            "Ensure this value has at most 100 characters (it has 101)."
-            in excinfo.value.message_dict['name']
-        )
-
